refactor: update path from di to dependency_injection
