+++ lib/brick_oven.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 
-import 'package:brick_oven/utils/di.dart';
+import 'package:brick_oven/utils/dependency_injection.dart';
 import 'package:file/file.dart';
 import 'package:file/local.dart';
 import 'package:mason_logger/mason_logger.dart';
