+++ lib/domain/brick_yaml_config.dart
@@ -1,5 +1,5 @@
 import 'package:autoequal/autoequal.dart';
-import 'package:brick_oven/utils/di.dart';
+import 'package:brick_oven/utils/dependency_injection.dart';
 import 'package:equatable/equatable.dart';
 import 'package:file/file.dart';
 import 'package:mason_logger/mason_logger.dart';
