+++ lib/domain/partial.dart
@@ -3,7 +3,7 @@ import 'package:brick_oven/domain/file_write_result.dart';
 import 'package:brick_oven/domain/variable.dart';
 import 'package:brick_oven/domain/yaml_value.dart';
 import 'package:brick_oven/src/exception.dart';
-import 'package:brick_oven/utils/di.dart';
+import 'package:brick_oven/utils/dependency_injection.dart';
 import 'package:brick_oven/utils/extensions/yaml_map_extensions.dart';
 import 'package:brick_oven/utils/file_replacements.dart';
 import 'package:equatable/equatable.dart';
