+++ lib/src/commands/brick_oven.dart
@@ -1,5 +1,5 @@
 import 'package:args/command_runner.dart';
-import 'package:brick_oven/utils/di.dart';
+import 'package:brick_oven/utils/dependency_injection.dart';
 import 'package:file/file.dart';
 import 'package:mason_logger/mason_logger.dart';
 import 'package:yaml/yaml.dart';
