+++ lib/utils/file_replacements.dart
@@ -1,4 +1,4 @@
-import 'package:brick_oven/utils/di.dart';
+import 'package:brick_oven/utils/dependency_injection.dart';
 import 'package:file/file.dart';
 import 'package:mason_logger/mason_logger.dart';
 import 'package:meta/meta.dart';
