+++ test/domain/brick_test.dart
@@ -2,7 +2,7 @@
 
 import 'dart:async';
 
-import 'package:brick_oven/utils/di.dart';
+import 'package:brick_oven/utils/dependency_injection.dart';
 import 'package:file/file.dart';
 import 'package:file/memory.dart';
 import 'package:mason_logger/mason_logger.dart';
