+++ test/domain/bricks_or_error_test.dart
@@ -1,7 +1,7 @@
 import 'package:brick_oven/domain/brick.dart';
 import 'package:brick_oven/domain/brick_source.dart';
 import 'package:brick_oven/domain/bricks_or_error.dart';
-import 'package:brick_oven/utils/di.dart';
+import 'package:brick_oven/utils/dependency_injection.dart';
 import 'package:mason_logger/mason_logger.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';
