+++ test/src/commands/update_test.dart
@@ -1,6 +1,6 @@
 import 'package:brick_oven/src/commands/update.dart';
 import 'package:brick_oven/src/version.dart';
-import 'package:brick_oven/utils/di.dart';
+import 'package:brick_oven/utils/dependency_injection.dart';
 import 'package:mason_logger/mason_logger.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:pub_updater/pub_updater.dart';
